Cosm message preview for mint templates
